refactor & fix(tests)
