+++ tests/package_managers/homebrew/test_homebrew_diff_deps.py
@@ -6,112 +6,13 @@ need to be added or removed when processing homebrew formula updates.
 """
 
 from datetime import datetime
-from uuid import uuid4
+from uuid import UUID, uuid4
 
 import pytest
 
 from core.models import LegacyDependency, Package
-from core.structs import Cache
 from package_managers.homebrew.diff import Diff
 from package_managers.homebrew.structs import Actual
-
-
-@pytest.fixture
-def package_ids():
-    """Fixture providing consistent package IDs for testing."""
-    return {"foo": uuid4(), "bar": uuid4(), "baz": uuid4(), "qux": uuid4()}
-
-
-@pytest.fixture
-def packages(package_ids):
-    """Fixture providing test packages."""
-    return {
-        "foo": Package(
-            id=package_ids["foo"],
-            name="foo",
-            package_manager_id=1,
-            import_id="foo",
-            created_at=datetime.now(),
-            updated_at=datetime.now(),
-        ),
-        "bar": Package(
-            id=package_ids["bar"],
-            name="bar",
-            package_manager_id=1,
-            import_id="bar",
-            created_at=datetime.now(),
-            updated_at=datetime.now(),
-        ),
-        "baz": Package(
-            id=package_ids["baz"],
-            name="baz",
-            package_manager_id=1,
-            import_id="baz",
-            created_at=datetime.now(),
-            updated_at=datetime.now(),
-        ),
-        "qux": Package(
-            id=package_ids["qux"],
-            name="qux",
-            package_manager_id=1,
-            import_id="qux",
-            created_at=datetime.now(),
-            updated_at=datetime.now(),
-        ),
-    }
-
-
-@pytest.fixture
-def diff_instance(mock_config):
-    """
-    Factory fixture to create Diff instances with specific cache configurations.
-
-    Returns a function that creates Diff instances.
-    """
-
-    def create_diff(package_map, dependencies=None, url_map=None, package_urls=None):
-        cache = Cache(
-            package_map=package_map,
-            url_map=url_map or {},
-            package_urls=package_urls or {},
-            dependencies=dependencies or {},
-        )
-        return Diff(mock_config, cache)
-
-    return create_diff
-
-
-@pytest.fixture
-def homebrew_formula():
-    """
-    Factory fixture to create Actual homebrew formula objects.
-
-    Returns a function that creates Actual objects.
-    """
-
-    def create_formula(
-        formula_name,
-        dependencies=None,
-        build_dependencies=None,
-        test_dependencies=None,
-        recommended_dependencies=None,
-        optional_dependencies=None,
-    ):
-        return Actual(
-            formula=formula_name,
-            description="Test formula",
-            license="MIT",
-            homepage="",
-            source="",
-            repository="",
-            dependencies=dependencies or [],
-            build_dependencies=build_dependencies or [],
-            test_dependencies=test_dependencies or [],
-            recommended_dependencies=recommended_dependencies or [],
-            optional_dependencies=optional_dependencies or [],
-        )
-
-    return create_formula
 
 
 @pytest.mark.transformer
@@ -142,7 +43,12 @@ class TestDiffDeps:
         assert len(removed_deps) == 0, "No removed deps for new pkg"
 
     def test_existing_package_adding_dependency(
-        self, packages, package_ids, diff_instance, homebrew_formula, mock_config
+        self,
+        packages: dict[str, Package],
+        package_ids: dict[str, UUID],
+        diff_instance: Diff,
+        homebrew_formula: Actual,
+        mock_config,
     ):
         """Test diff_deps when adding a new dependency to an existing package."""
         # Create existing dependency
@@ -150,15 +56,20 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.runtime.id,
+            dependency_type_id=mock_config.dependency_types.runtime,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
 
         # Create diff with existing dependency
-        diff = diff_instance(
+        diff: Diff = diff_instance(
             package_map=packages, dependencies={package_ids["foo"]: {existing_dep}}
         )
+        for k, s in diff.caches.dependencies.items():
+            print(f"Package: {k}")
+            for v in s:
+                print(v.dependency_id)
+                print(v.dependency_type_id)
 
         # Create formula with existing dependency plus a new one
         pkg = homebrew_formula(
@@ -170,6 +81,11 @@ class TestDiffDeps:
         # Execute
         new_deps, removed_deps = diff.diff_deps(pkg)
 
+        for dep in new_deps:
+            print(
+                f"{dep.package_id} -> {dep.dependency_id} as {dep.dependency_type_id}"
+            )
+
         # Assert
         assert len(new_deps) == 1, "One new dependency should be added"
         assert len(removed_deps) == 0, "No dependencies should be removed"
@@ -178,7 +94,7 @@ class TestDiffDeps:
         new_dep = new_deps[0]
         assert new_dep.package_id == package_ids["foo"]
         assert new_dep.dependency_id == package_ids["baz"]
-        assert new_dep.dependency_type_id == mock_config.dependency_types.build.id
+        assert new_dep.dependency_type_id == mock_config.dependency_types.build
 
     def test_existing_package_removing_dependency(
         self, packages, package_ids, diff_instance, homebrew_formula, mock_config
@@ -189,7 +105,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.runtime.id,
+            dependency_type_id=mock_config.dependency_types.runtime,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -198,7 +114,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["baz"],
-            dependency_type_id=mock_config.dependency_types.build.id,
+            dependency_type_id=mock_config.dependency_types.build,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -226,7 +142,7 @@ class TestDiffDeps:
         removed_dep = removed_deps[0]
         assert removed_dep.package_id == package_ids["foo"]
         assert removed_dep.dependency_id == package_ids["baz"]
-        assert removed_dep.dependency_type_id == mock_config.dependency_types.build.id
+        assert removed_dep.dependency_type_id == mock_config.dependency_types.build
 
     def test_existing_package_changing_dependency_type(
         self, packages, package_ids, diff_instance, homebrew_formula, mock_config
@@ -240,7 +156,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.runtime.id,
+            dependency_type_id=mock_config.dependency_types.runtime,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -268,13 +184,13 @@ class TestDiffDeps:
         removed_dep = removed_deps[0]
         assert removed_dep.package_id == package_ids["foo"]
         assert removed_dep.dependency_id == package_ids["bar"]
-        assert removed_dep.dependency_type_id == mock_config.dependency_types.runtime.id
+        assert removed_dep.dependency_type_id == mock_config.dependency_types.runtime
 
         # Verify new dep is build
         new_dep = new_deps[0]
         assert new_dep.package_id == package_ids["foo"]
         assert new_dep.dependency_id == package_ids["bar"]
-        assert new_dep.dependency_type_id == mock_config.dependency_types.build.id
+        assert new_dep.dependency_type_id == mock_config.dependency_types.build
 
     def test_existing_package_no_dependency_changes(
         self, packages, package_ids, diff_instance, homebrew_formula, mock_config
@@ -288,7 +204,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.runtime.id,
+            dependency_type_id=mock_config.dependency_types.runtime,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -326,7 +242,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.runtime.id,
+            dependency_type_id=mock_config.dependency_types.runtime,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -362,7 +278,7 @@ class TestDiffDeps:
             id=uuid4(),
             package_id=package_ids["foo"],
             dependency_id=package_ids["bar"],
-            dependency_type_id=mock_config.dependency_types.build.id,
+            dependency_type_id=mock_config.dependency_types.build,
             created_at=datetime.now(),
             updated_at=datetime.now(),
         )
@@ -385,11 +301,11 @@ class TestDiffDeps:
 
         # Assert
         assert len(new_deps) == 1, "One new dependency should be added"
-        assert (
-            new_deps[0].dependency_type_id == mock_config.dependency_types.runtime.id
-        ), "The new dependency should be runtime"
+        assert new_deps[0].dependency_type_id == mock_config.dependency_types.runtime, (
+            "The new dependency should be runtime"
+        )
 
         assert len(removed_deps) == 1, "The build dependency should be removed"
         assert (
-            removed_deps[0].dependency_type_id == mock_config.dependency_types.build.id
+            removed_deps[0].dependency_type_id == mock_config.dependency_types.build
         ), "The removed dependency should be build"
